project-semantic-search/server/requirements.txt
Flask==2.2.5
pandas
mysql-connector-python
pymysql
flask-cors
numpy
openai==0.28
python-dotenv